+++ src/main/java/com/app/model/Invoice.java
@@ -15,23 +15,23 @@ import jakarta.persistence.Table;
 public class Invoice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer code;
+    private Long code;
 
     private Double cost;
     private Double paid;
     private String description;
 
     @Column(name = "po_id")
-    private Integer poId;
+    private Long poId;
 
     @Column(name = "job_id")
-    private Integer jobId;
+    private Long jobId;
 
     @Column(name = "by_id")
-    private Integer byId;
+    private Long byId;
 
     @Column(name = "project_id")
-    private Integer projectId;
+    private Long projectId;
 
     private String ref;
 
@@ -49,8 +49,8 @@ public class Invoice {
     private OffsetDateTime updatedAt;
 
     // Getters and Setters
-    public Integer getCode() { return code; }
-    public void setCode(Integer code) { this.code = code; }
+    public Long getCode() { return code; }
+    public void setCode(Long code) { this.code = code; }
 
     public Double getCost() { return cost; }
     public void setCost(Double cost) { this.cost = cost; }
@@ -61,17 +61,17 @@ public class Invoice {
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
 
-    public Integer getPoId() { return poId; }
-    public void setPoId(Integer poId) { this.poId = poId; }
+    public Long getPoId() { return poId; }
+    public void setPoId(Long poId) { this.poId = poId; }
 
-    public Integer getJobId() { return jobId; }
-    public void setJobId(Integer jobId) { this.jobId = jobId; }
+    public Long getJobId() { return jobId; }
+    public void setJobId(Long jobId) { this.jobId = jobId; }
 
-    public Integer getById() { return byId; }
-    public void setById(Integer byId) { this.byId = byId; }
+    public Long getById() { return byId; }
+    public void setById(Long byId) { this.byId = byId; }
 
-    public Integer getProjectId() { return projectId; }
-    public void setProjectId(Integer projectId) { this.projectId = projectId; }
+    public Long getProjectId() { return projectId; }
+    public void setProjectId(Long projectId) { this.projectId = projectId; }
 
     public String getRef() { return ref; }
     public void setRef(String ref) { this.ref = ref; }
